refactor utils/files to meet codeclimate compliance
