+++ utils/files/newFile.js
@@ -1,0 +1,15 @@
+const FileAPI = require('./FileAPI')
+
+/**
+ * New File
+ *
+ * Creates an empty File object
+ *
+ * @param {string} filename - the filename without path info
+ */
+function newFile(filename) {
+  var File = FileAPI()
+  return new File([''], filename)
+}
+
+module.exports = newFile
